refactor: Consolidate and enhance helper functions across modules
- Introduced centralized formatHelper for consistent data formatting (currency, date, number).
- Updated authHelper to include error handling for authentication processes.
- Removed redundant helper files and streamlined imports in various modules.
- Enhanced metadata generation for forms and search functionalities across multiple modules.
- Improved code readability and maintainability by consolidating similar functionalities.
