+++ src/helpers/routesHelper.ts
@@ -1,44 +1,97 @@
-// Rutas principales del sistema
+/**
+ * Constantes de rutas del sistema de pasantías
+ *
+ * Centraliza todas las rutas de la aplicación para evitar hardcoding
+ * y facilitar el mantenimiento y refactoring.
+ */
 export const ROUTES = {
-  // Rutas públicas
+  // ==================== RUTAS PÚBLICAS ====================
+  /** Página de inicio de sesión */
   LOGIN: '/login',
+  /** Página de registro de usuarios */
   REGISTRAR_USUARIO: '/register',
 
-  // Rutas protegidas
+  // ==================== RUTAS PROTEGIDAS ====================
+  /** Dashboard principal del sistema */
   DASHBOARD: '/dashboard',
-  INICIO: '/dashboard', // Alias para dashboard
 
-  // Módulos principales
+  // ==================== MÓDULO EMPRESAS ====================
+  /** Listado de empresas */
   EMPRESAS: '/empresas',
+  /** Crear nueva empresa */
   EMPRESAS_CREAR: '/empresas/crear',
+  /** Editar empresa existente */
   EMPRESAS_EDITAR: '/empresas/editar',
+  /** Ver detalle de empresa */
   EMPRESAS_DETALLE: '/empresas/detalle',
+
+  // ==================== MÓDULO ESTUDIANTES ====================
+  /** Listado de estudiantes */
   ESTUDIANTES: '/students',
+  /** Crear nuevo estudiante */
   ESTUDIANTES_CREAR: '/students/crear',
+  /** Editar estudiante existente */
   ESTUDIANTES_EDITAR: '/students/editar',
+  /** Ver detalle de estudiante */
   ESTUDIANTES_DETALLE: '/students/detalle',
+
+  // ==================== MÓDULO CONVENIOS ====================
+  /** Listado de convenios */
   CONVENIOS: '/convenios',
+  /** Crear nuevo convenio */
   CONVENIOS_CREAR: '/convenios/crear',
+  /** Editar convenio existente */
   CONVENIOS_EDITAR: '/convenios/editar',
+  /** Ver detalle de convenio */
   CONVENIOS_DETALLE: '/convenios/detalle',
+
+  // ==================== MÓDULO PASANTÍAS ====================
+  /** Listado de pasantías */
   PASANTIAS: '/pasantias',
+  /** Crear nueva pasantía */
   PASANTIAS_CREAR: '/pasantias/crear',
+  /** Editar pasantía existente */
   PASANTIAS_EDITAR: '/pasantias/editar',
+  /** Ver detalle de pasantía */
   PASANTIAS_DETALLE: '/pasantias/detalle',
+
+  // ==================== MÓDULO PAGOS ====================
+  /** Listado de pagos */
   PAGOS: '/pagos',
+  /** Crear nuevo pago */
   PAGOS_CREAR: '/pagos/crear',
+  /** Editar pago existente */
   PAGOS_EDITAR: '/pagos/editar',
+  /** Ver detalle de pago */
   PAGOS_DETALLE: '/pagos/detalle',
+
+  // ==================== MÓDULOS ADICIONALES ====================
+  /** Página de reportes */
   REPORTES: '/reportes',
 
-  // Usuario y configuración
+  // ==================== USUARIO Y CONFIGURACIÓN ====================
+  /** Perfil del usuario */
   PERFIL: '/perfil',
+  /** Configuración del sistema */
   CONFIGURACION: '/configuracion',
 
-  // Página de error
+  // ==================== PÁGINAS DE ERROR ====================
+  /** Página 404 - No encontrado */
   NOT_FOUND: '*',
 } as const;
 
-// Tipo para las rutas (opcional, para TypeScript)
+// ==================== TIPOS TYPESCRIPT ====================
+
+/** Tipo que representa las claves de las rutas */
 export type RouteKey = keyof typeof ROUTES;
+
+/** Tipo que representa los valores de las rutas */
 export type RouteValue = (typeof ROUTES)[RouteKey];
+
+// ==================== UTILIDADES DE RUTAS ====================
+
+/**
+ * Lista de rutas públicas que no requieren autenticación
+ * Utilizada por ProtectedRoute para determinar acceso
+ */
+export const PUBLIC_ROUTES = [ROUTES.LOGIN, ROUTES.REGISTRAR_USUARIO] as const;
