+++ src/lib/FormularioGenerico/index.ts
@@ -22,21 +22,5 @@ export type {
   GenericFormProps,
 } from './types';
 
-// Helpers
-export {
-  createEmailValidation,
-  createRequiredValidation,
-  createMinLengthValidation,
-  createMaxLengthValidation,
-  createNumberRangeValidation,
-} from './helpers/validationHelpers';
-
-export {
-  createDropdownOptions,
-  createTextField,
-  createEmailField,
-  createDateField,
-  createCurrencyField,
-  createDropdownField,
-  createDynamicDropdownField,
-} from './helpers/formMetadataHelpers';
+// Nota: Las validaciones se definen directamente inline en cada helper de módulo
+// No se necesitan funciones helper adicionales para validaciones
