+++ src/modules/Empresas/index.ts
@@ -2,8 +2,11 @@ export { default as Empresas } from './Empresas';
 export { default as CreacionEmpresas } from './CreacionEmpresas';
 export { default as EditarEmpresa } from './EditarEmpresa';
 export type { EmpresaDto, CreacionEmpresaDto, EmpresasState } from './types';
-export { getEmpresasStats } from './helpers/empresaHelpers';
 export {
+  getEmpresasStats,
   getCreacionEmpresaMetadata,
   getEdicionEmpresaMetadata,
-} from './helpers/creacionEmpresaHelpers';
+  getEmpresaSearchMetadata,
+  formatEmpresaSearchFilters,
+  getSugerenciasNombresEmpresas,
+} from './helpers/empresaHelpers';
