+++ src/modules/Pasantias/helpers/pasantiaHelpers.ts
@@ -1,6 +1,32 @@
-import { PasantiaStats, PasantiaDto, PasantiaFilters } from '../types';
+import {
+  PasantiaStats,
+  PasantiaDto,
+  PasantiaBusquedaAvanzadaDto,
+} from '../types';
+import {
+  TIPOS_ACUERDO_VALIDOS,
+  CARRERAS_VALIDAS,
+} from '../../../helpers/constants';
+import { FieldMetadata } from '../../../lib/ElementCardGenerica';
 
-// Metadata para formularios de pasantías
+/**
+ * Helper consolidado para pasantías
+ *
+ * Contiene funciones utilitarias para:
+ * - Generación de metadata de formularios (crear/editar)
+ * - Generación de metadata de formularios de búsqueda
+ * - Formateo de filtros de búsqueda para el backend
+ * - Cálculo de estadísticas de pasantías
+ */
+
+// ==================== METADATA DE FORMULARIOS ====================
+
+/**
+ * Genera la metadata para el formulario de pasantías
+ * Utilizado en CrearPasantia y EditarPasantia
+ *
+ * @returns Configuración completa del formulario con campos, validaciones y opciones
+ */
 export const getPasantiaFormMetadata = () => ({
   title: 'Información de la Pasantía',
   submitButtonText: 'Guardar',
@@ -125,11 +151,7 @@ export const getPasantiaFormMetadata = () => ({
       name: 'tipoAcuerdo',
       label: 'Tipo de Acuerdo',
       type: 'dropdown' as const,
-      options: [
-        { value: 'Pasantia', label: 'Pasantía' },
-        { value: 'PPS', label: 'PPS' },
-        { value: 'otro', label: 'Otro' },
-      ],
+      options: TIPOS_ACUERDO_VALIDOS,
       gridSize: 6,
     },
     {
@@ -175,7 +197,15 @@ export const getPasantiaFormMetadata = () => ({
   ],
 });
 
-// Función para calcular estadísticas de pasantías
+// ==================== CÁLCULO DE ESTADÍSTICAS ====================
+
+/**
+ * Calcula las estadísticas de pasantías basadas en un array de pasantías
+ *
+ * @param pasantias - Array de pasantías para calcular estadísticas
+ * @param pasantiasPorVencer - Número opcional de pasantías por vencer (si no se proporciona, se calcula automáticamente)
+ * @returns Objeto con estadísticas calculadas (total, activas, finalizadas, por vencer)
+ */
 export const calculatePasantiaStats = (
   pasantias: PasantiaDto[],
   pasantiasPorVencer?: number
@@ -217,152 +247,111 @@ export const calculatePasantiaStats = (
   };
 };
 
-// Función para filtrar pasantías
-export const filterPasantias = (
-  pasantias: PasantiaDto[],
-  filters: PasantiaFilters
-): PasantiaDto[] => {
-  return pasantias.filter(pasantia => {
-    // Filtro por trámite
-    if (
-      filters.expediente &&
-      !pasantia.tramite
-        ?.toLowerCase()
-        .includes(filters.expediente.toLowerCase())
-    ) {
-      return false;
-    }
+// ==================== METADATA DE BÚSQUEDA ====================
 
-    // Filtro por obra social
-    if (
-      filters.obraSocial &&
-      !pasantia.obraSocial
-        ?.toLowerCase()
-        .includes(filters.obraSocial.toLowerCase())
-    ) {
-      return false;
-    }
-
-    // Filtro por ART
-    if (
-      filters.art &&
-      !pasantia.art?.toLowerCase().includes(filters.art.toLowerCase())
-    ) {
-      return false;
-    }
-
-    // Filtro por tutor empresa
-    if (
-      filters.tutorEmpresa &&
-      !pasantia.tutorEmpresa
-        ?.toLowerCase()
-        .includes(filters.tutorEmpresa.toLowerCase())
-    ) {
-      return false;
-    }
-
-    // Filtro por tutor facultad
-    if (
-      filters.tutorFacultad &&
-      !pasantia.tutorFacultad
-        ?.toLowerCase()
-        .includes(filters.tutorFacultad.toLowerCase())
-    ) {
-      return false;
-    }
-
-    // Filtro por tipo de acuerdo
-    if (filters.tipoAcuerdo && pasantia.tipoAcuerdo !== filters.tipoAcuerdo) {
-      return false;
-    }
-
-    // Filtro por fecha de inicio
-    if (filters.fechaInicioDesde && pasantia.fechaInicio) {
-      const fechaInicio = new Date(pasantia.fechaInicio);
-      const fechaDesde = new Date(filters.fechaInicioDesde);
-      if (fechaInicio < fechaDesde) return false;
-    }
-
-    if (filters.fechaInicioHasta && pasantia.fechaInicio) {
-      const fechaInicio = new Date(pasantia.fechaInicio);
-      const fechaHasta = new Date(filters.fechaInicioHasta);
-      if (fechaInicio > fechaHasta) return false;
-    }
-
-    // Filtro por fecha de fin
-    if (filters.fechaFinDesde && pasantia.fechaFin) {
-      const fechaFin = new Date(pasantia.fechaFin);
-      const fechaDesde = new Date(filters.fechaFinDesde);
-      if (fechaFin < fechaDesde) return false;
-    }
-
-    if (filters.fechaFinHasta && pasantia.fechaFin) {
-      const fechaFin = new Date(pasantia.fechaFin);
-      const fechaHasta = new Date(filters.fechaFinHasta);
-      if (fechaFin > fechaHasta) return false;
-    }
-
-    return true;
-  });
-};
-
-// Función para formatear fechas
-export const formatDate = (dateString?: string): string => {
-  if (!dateString) return 'No especificada';
-  const date = new Date(dateString);
-  return date.toLocaleDateString('es-ES', {
-    year: 'numeric',
-    month: 'long',
-    day: 'numeric',
-  });
-};
-
-// Función para verificar si una pasantía está activa
-export const isPasantiaActiva = (fechaFin?: string): boolean => {
-  if (!fechaFin) return false;
-  const fechaActual = new Date();
-  const fechaFinDate = new Date(fechaFin);
-  return fechaFinDate > fechaActual;
-};
-
-// Función para obtener el estado de una pasantía
-export const getPasantiaEstado = (fechaFin?: string): string => {
-  if (!fechaFin) return 'Sin fecha de fin';
-
-  const fechaActual = new Date();
-  const fechaFinDate = new Date(fechaFin);
-
-  if (fechaFinDate <= fechaActual) {
-    return 'Finalizada';
-  }
-
-  const diasRestantes = Math.ceil(
-    (fechaFinDate.getTime() - fechaActual.getTime()) / (1000 * 60 * 60 * 24)
-  );
-
-  if (diasRestantes <= 30) {
-    return `Por vencer (${diasRestantes} días)`;
-  }
-  return 'Activa';
-};
-
-// Función para obtener valores por defecto del formulario
-export const getDefaultPasantiaValues = () => ({
-  idEstudiante: undefined,
-  idConvenio: undefined,
-  asignacionMensual: 0,
-  obraSocial: '',
-  art: '',
-  tutorEmpresa: '',
-  tutorFacultad: '',
-  dniTutorFacultad: '',
-  fechaInicio: '',
-  fechaFin: '',
-  tipoAcuerdo: 'Pasantia',
-  frecuenciaPago: 'Mensual',
-  montoPago: 0,
-  observaciones: '',
-  areaTrabajo: '',
-  estado: 'Activa',
-  sudocu: '',
+/**
+ * Genera la metadata para el formulario de búsqueda avanzada de pasantías
+ * Utilizado en PasantiaFilters
+ *
+ * @returns Configuración completa del formulario de búsqueda con campos y opciones
+ */
+export const getPasantiaSearchMetadata = (): {
+  title: string;
+  fields: FieldMetadata[];
+  submitButtonText: string;
+  cancelButtonText: string;
+} => ({
+  title: 'Búsqueda Avanzada de Pasantías',
+  fields: [
+    {
+      name: 'numeroTramite',
+      label: 'Número de Trámite',
+      type: 'dynamicDropdown',
+      placeholder: 'Seleccionar número de trámite...',
+    },
+    {
+      name: 'tipo',
+      label: 'Tipo',
+      type: 'dropdown',
+      options: [{ value: '', label: 'Todos' }, ...TIPOS_ACUERDO_VALIDOS],
+    },
+    {
+      name: 'estudiante',
+      label: 'Documento del Estudiante',
+      type: 'dynamicDropdown',
+      placeholder: 'Seleccionar documento del estudiante...',
+    },
+    {
+      name: 'empresa',
+      label: 'Empresa',
+      type: 'dynamicDropdown',
+      placeholder: 'Seleccionar empresa...',
+    },
+    {
+      name: 'vigente',
+      label: 'Vigente',
+      type: 'dropdown',
+      options: [
+        { value: '', label: 'Todos' },
+        { value: 'vigente', label: 'Vigente' },
+        { value: 'no_vigente', label: 'No Vigente' },
+      ],
+    },
+    {
+      name: 'carrera',
+      label: 'Carrera',
+      type: 'dropdown',
+      options: [
+        { value: '', label: 'Todas las carreras' },
+        ...CARRERAS_VALIDAS.map(carrera => ({
+          value: carrera,
+          label: carrera,
+        })),
+      ],
+    },
+  ],
+  submitButtonText: 'Buscar',
+  cancelButtonText: 'Limpiar',
 });
+
+// ==================== FORMATEO DE FILTROS ====================
+
+/**
+ * Formatea los filtros del formulario de búsqueda para enviarlos al backend
+ * Convierte los valores del frontend al formato esperado por la API
+ *
+ * @param filters - Filtros del formulario de búsqueda
+ * @returns Objeto DTO formateado para el backend
+ */
+export const formatPasantiaSearchFilters = (
+  filters: Record<string, unknown>
+): PasantiaBusquedaAvanzadaDto => {
+  const formattedFilters: PasantiaBusquedaAvanzadaDto = {};
+
+  if (filters.numeroTramite) {
+    formattedFilters.numeroTramite = filters.numeroTramite as string;
+  }
+  if (filters.tipo) {
+    formattedFilters.tipo = filters.tipo as string;
+  }
+  if (filters.estudiante) {
+    formattedFilters.estudiante = filters.estudiante as string;
+  }
+  if (filters.empresa) {
+    formattedFilters.empresa = filters.empresa as string;
+  }
+  if (filters.vigente) {
+    // Convertir string del frontend a boolean esperado por el backend
+    if (filters.vigente === 'vigente') {
+      formattedFilters.vigente = true;
+    } else if (filters.vigente === 'no_vigente') {
+      formattedFilters.vigente = false;
+    }
+    // Si es empty string o 'todos', no se incluye el filtro
+  }
+  if (filters.carrera) {
+    formattedFilters.carrera = filters.carrera as string;
+  }
+
+  return formattedFilters;
+};
