+++ src/modules/Shared/components/header/PagosVencerNotifications.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Payment } from '@mui/icons-material';
+import { formatCurrency } from '../../../../helpers/formatHelper';
 import { usePagosPorVencer } from '../../../Pagos/hooks/usePagos';
 import { PagosDto } from '../../../Pagos/types';
 import VencimientosNotifications, {
@@ -30,7 +31,7 @@ const PagosVencerNotifications: React.FC<PagosVencerNotificationsProps> = ({
     },
     getItemSubtitle: (item: VencimientoItem) => {
       const pago = item as unknown as PagosDto;
-      return `Monto: $${pago.monto?.toLocaleString() || 'N/A'}`;
+      return `Monto: ${formatCurrency(pago.monto) || 'N/A'}`;
     },
     onItemClick: (item: VencimientoItem) => {
       onNavigateToPago?.(item.id);
